services/baxus-monitor/requirements.txt
google-cloud-pubsub>=2.18.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
heliuspy
solana
google-generativeai
